Lint fix: printf-like formatting function 'log' should be named 'logf'

+++ src/remote/project.go
@@ -67,13 +67,13 @@ func newTestProject(t *testing.T, api *StorageApi, stateFilePath string) *testPr
 
 	// Create
 	p := &testProject{t, &sync.Mutex{}, testDir, stateFile, api, defaultBranch, nil}
-	p.log("Initializing test project \"%s\", id: \"%d\".", p.api.ProjectName(), p.api.ProjectId())
+	p.logf("Initializing test project \"%s\", id: \"%d\".", p.api.ProjectName(), p.api.ProjectId())
 	return p
 }
 
 // Clear deletes all project branches (except default) and all configurations.
 func (p *testProject) Clear() {
-	p.log("Clearing project ...")
+	p.logf("Clearing project ...")
 	startTime := time.Now()
 
 	// Delete all configs in default branch, it cannot be deleted
@@ -96,13 +96,13 @@ func (p *testProject) Clear() {
 		}
 	}
 
-	p.log("Test project cleared | %s", time.Since(startTime))
+	p.logf("Test project cleared | %s", time.Since(startTime))
 }
 
 // InitState creates branches and configurations according stateFile.
 func (p *testProject) InitState() {
 	startTime := time.Now()
-	p.log("Setting project state ...")
+	p.logf("Setting project state ...")
 
 	// Create configs in default branch, they will be auto-copied to dev-branches
 	pool := p.api.NewPool()
@@ -126,7 +126,7 @@ func (p *testProject) InitState() {
 			p.api.
 				CreateBranchRequest(branch).
 				OnSuccess(func(response *client.Response) {
-					p.log(`crated branch "%s", id: "%d"`, branch.Name, branch.Id)
+					p.logf(`crated branch "%s", id: "%d"`, branch.Name, branch.Id)
 					p.setEnv(fmt.Sprintf("TEST_BRANCH_%s_ID", branch.Name), cast.ToString(branch.Id))
 				}).
 				Send()
@@ -146,11 +146,11 @@ func (p *testProject) InitState() {
 
 	// Log ENVs
 	for _, env := range p.envs {
-		p.log(fmt.Sprintf(`Set ENV "%s"`, env))
+		p.logf(fmt.Sprintf(`Set ENV "%s"`, env))
 	}
 
 	// Done
-	p.log("Project state set | %s", time.Since(startTime))
+	p.logf("Project state set | %s", time.Since(startTime))
 }
 
 // CreateConfigsInBranch loads configs from files and creates them in the test project.
@@ -159,7 +159,7 @@ func (p *testProject) CreateConfigsInBranch(pool *client.Pool, names []string, b
 		config := fixtures.LoadConfig(p.t, name)
 		config.BranchId = branch.Id
 		if request, err := p.api.CreateConfigRequest(config); err == nil {
-			p.log("creating config \"%s/%s/%s\"", branch.Name, config.ComponentId, config.Name)
+			p.logf("creating config \"%s/%s/%s\"", branch.Name, config.ComponentId, config.Name)
 			pool.
 				Request(request).
 				OnSuccess(func(response *client.Response) {
@@ -191,7 +191,7 @@ func (p *testProject) setEnv(key string, value string) {
 	p.envs = append(p.envs, fmt.Sprintf("%s=%s", key, value))
 }
 
-func (p *testProject) log(format string, a ...interface{}) {
+func (p *testProject) logf(format string, a ...interface{}) {
 	if utils.TestIsVerbose() {
 		a = append([]interface{}{p.t.Name()}, a...)
 		fmt.Println(fmt.Sprintf("Fixtures[%s]: "+format, a...))
